refactor(canary): split SimulationCanaryService into partial-class files (no behavior change)
Mechanically extracts the 1081-LoC SimulationCanaryService.cs into 6 partial-class
files grouped by responsibility. The compiled type, DI registration, field
initialization order, and public/internal method surface are unchanged —
partial classes compile to a single type with byte-identical semantics.

Same pattern as PR #428 (CirclesRpcModule split) and PR #429 (CacheWarmupService split).

File layout in src/Pathfinder/Circles.Pathfinder.Host/Canary/:
- SimulationCanaryService.cs                       (134 LoC) — fields, ctor, TryEnqueue, metrics, CanaryWorkItem record
- SimulationCanaryService.Lifecycle.cs             (182)     — ExecuteAsync, SimulateAsync (eth_call path)
- SimulationCanaryService.UnwrapCalldata.cs        (194)     — Demurraged/Resolved records, BuildUnwrapPrefix, EncodeUnwrapCalldata, PromoteAllDemurraged
- SimulationCanaryService.BundleSimulation.cs      (289)     — SimulateBundleAsync, BundleOutcome/BundleParseResult, ParseSimulateV1Response, ResolveWrapperTokenOwners
- SimulationCanaryService.InflationaryUnits.cs     (203)     — ComputeInflationDay, EncodeConvertDemurrageCalldata, ParseConvertCallReturnData, ExtractInflationaryAmounts, ApplyInflationaryAmounts, ApplyInflationaryRoundtripBump
- SimulationCanaryService.InflationaryResolver.cs  (246)     — FetchBlockTimestampAsync, ResolveInflationaryAmountsAsync

Verification (baseline captured from origin/staging before any edit):
- Pathfinder.Host clean build: 0 errors, 20 warnings (identical to baseline)
- Full repo build: 0 errors (clean)
- Pathfinder.Tests: 1091 pass / 0 fail / 270 skip / 1361 total (identical to baseline)
- Source method body equivalence: 18 bodies in baseline, 18 in post-split, all 18 byte-identical
  after whitespace normalization (zero differences, none added, none removed)

Out of scope: any behavior change, type renames, the helper-class extraction
pattern suggested by the earlier canary follow-up memory. The encapsulation
goals from that memory remain valid but are deferred to a separate refactor.
